fix(clippy): resolve --features isal-compression lint failures
The CI Lint job runs `cargo clippy --release --features isal-compression`,
which compiles the x86_64-gated parallel-SM code a default-features build
skips. 14 lint errors there were blocking CI:

- unnecessary_cast, needless_borrow, unused_assignments, collapsible_match
  fixed directly.
- while_let_loop (loop is a faithful port of vendor processNextChunk),
  needless_late_init, large_enum_variant (boxing ChunkData would add a
  per-chunk alloc on the write path), and dead_code (diagnostic helpers
  kept for future routing) suppressed with a justified #[allow].

+++ src/decompress/parallel/deflate_block.rs
@@ -992,7 +992,7 @@ impl Block {
                 if code <= 255 || sym_count > 1 {
                     // SAFETY: see ring_ptr SAFETY note above.
                     unsafe {
-                        ring_ptr.add(pos % RING_SIZE).write((code & 0xFF) as u16);
+                        ring_ptr.add(pos % RING_SIZE).write(code & 0xFF);
                     }
                     pos += 1;
                     emitted += 1;
